dist: missing glib-compile-resources in make dist
The *glib-compile-resources* is created durring the *make all* compiled, and
this binary also used to generate some source files. That dependency is not
expressed, but just works because of ordering (god forbid -j N option). I did
explored expressing such dependencies in autotool/automake for the *make dist*
target, but could not find a proper solution. Older glib also provided a way to
use the already compiled *glib-compile-resources*, that has been removed. Based
on that - without clear guide in glib itself - I concluded that *make dist*
requires a prior *make all*. This can be achieved with the *prepare_dist_hook*
setting as *make all*. The *prepare_dist_hook* is called between a *configure*
and *create_tarball* step.

Signed-off-by: Kokan <kokaipeter@gmail.com>