dockerfile: obtain the required archives prior the docker build command
Trigger for this change: ubuntu:lucid has openssl 0.9.8k which do not
support TLSv1.2. Because of this kernel.org terminated the connection
with SSL error while trying to download git tar.gz with wget.

Instead of constantly searching for new mirrors because of the old
openssl version, we download the archives on the host machine.

Signed-off-by: Laszlo Szemere <laszlo.szemere@oneidentity.com>